Add back, refresh and search. Use dense table layout and right align headers
