+++ azure-storage-web-client/src/main.js
@@ -6,17 +6,7 @@ import vuetify from './plugins/vuetify'
 Vue.config.productionTip = false
 
 Vue.use(VueRouter)
-
-const Foo = { template: '<div>foo</div>' }
-const Bar = { template: '<div>bar</div>' }
-const routes = [
-  { path: '/foo', component: Foo },
-  { path: '/bar', component: Bar }
-]
-const router = new VueRouter({
-  routes // short for `routes: routes`
-})
-
+const router = new VueRouter()
 
 new Vue({
   router,
